add separate command function module
